requirements.txt
httpx==0.26.0
schedule==1.2.1
python-dotenv==1.0.0
